+++ NEWS.md
@@ -5,6 +5,7 @@
     * Functions like `ma_article`, which return article content from a terminal endpoint are not memoized.
 * Added `ma_search` function for using the Memory Alpha site search via URL parameters.
 * Added `ma_image` function for downloading source images from Memory Alpha and loading into R as ggplot objects.
+* Added an initial version of an API function `memory_beta` and other `mb_*` Memory Beta analogs to Memory Alpha website functions.
 * Added new datasets, mainly to timelines: `tlBooks`, `tlEvents`, `tlFootnotes`. Also added `stSeries` names and abbreviations table.
 * Renamed `st_book_series` and `stBooksWP` to `st_books_wiki` and `stBooksWiki`, respectively.
 * Updated tile set URLs.